feat(skills): bound collector retry delays
## What

- Centralize shared external-source access for the diagnostic and repository-health collectors.
- Bound `Retry-After` delays to five seconds so transient rate-limit responses cannot stall collection indefinitely.

## Why

- Sharing the source-access boundary removes duplicated network and command helpers while keeping both documented collector commands on the same behavior.
- A bounded retry delay preserves responsive, read-only diagnostics when an upstream service returns a large rate-limit value.

## Testing

- `make dep` - passed
- `make clean-dep` - passed
- `make scripts-lint` - passed
- `make skills-lint` - passed
- `make docker-lint` - passed
- `make lint` - passed
- `make sec` - passed
- No maintained collector runtime fixture harness exists; the user explicitly approved no new tests for this fix.

AI-assisted-by: [Codex](https://openai.com/codex/)
